Airflow/requirements.txt
boto3==1.28.58
requests==2.31.0
python-dotenv==1.0.0
selenium==4.12.0
snowflake-connector-python==3.12.3
